+++ src/test/kotlin/com/fwdekker/randomness/datetime/DateTimeSchemeEditorTest.kt
@@ -1,22 +1,20 @@
 package com.fwdekker.randomness.datetime
 
-import com.fwdekker.randomness.editorApplyTestFactory
-import com.fwdekker.randomness.editorFieldsTestFactory
 import com.fwdekker.randomness.testhelpers.Tags
 import com.fwdekker.randomness.testhelpers.afterNonContainer
 import com.fwdekker.randomness.testhelpers.beforeNonContainer
 import com.fwdekker.randomness.testhelpers.dateTimeProp
+import com.fwdekker.randomness.testhelpers.editorApplyTests
+import com.fwdekker.randomness.testhelpers.editorFieldsTests
 import com.fwdekker.randomness.testhelpers.guiGet
 import com.fwdekker.randomness.testhelpers.guiRun
+import com.fwdekker.randomness.testhelpers.installEdtViolationDetection
 import com.fwdekker.randomness.testhelpers.prop
 import com.fwdekker.randomness.testhelpers.textProp
 import com.fwdekker.randomness.testhelpers.valueProp
-import com.intellij.testFramework.fixtures.IdeaTestFixture
-import com.intellij.testFramework.fixtures.IdeaTestFixtureFactory
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
-import org.assertj.swing.edt.FailOnThreadViolationRepaintManager
 import org.assertj.swing.fixture.Containers
 import org.assertj.swing.fixture.FrameFixture
 
@@ -25,27 +23,17 @@ import org.assertj.swing.fixture.FrameFixture
  * Unit tests for [DateTimeSchemeEditor].
  */
 object DateTimeSchemeEditorTest : FunSpec({
-    tags(Tags.EDITOR, Tags.IDEA_FIXTURE, Tags.SWING)
+    tags(Tags.EDITOR)
 
 
-    lateinit var ideaFixture: IdeaTestFixture
     lateinit var frame: FrameFixture
     lateinit var scheme: DateTimeScheme
     lateinit var editor: DateTimeSchemeEditor
 
 
-    beforeSpec {
-        FailOnThreadViolationRepaintManager.install()
-    }
-
-    afterSpec {
-        FailOnThreadViolationRepaintManager.uninstall()
-    }
+    installEdtViolationDetection()
 
     beforeNonContainer {
-        ideaFixture = IdeaTestFixtureFactory.getFixtureFactory().createBareFixture()
-        ideaFixture.setUp()
-
         scheme = DateTimeScheme()
         editor = guiGet { DateTimeSchemeEditor(scheme) }
         frame = Containers.showInFrame(editor.rootComponent)
@@ -53,7 +41,6 @@ object DateTimeSchemeEditorTest : FunSpec({
 
     afterNonContainer {
         frame.cleanUp()
-        ideaFixture.tearDown()
     }
 
 
@@ -67,10 +54,10 @@ object DateTimeSchemeEditorTest : FunSpec({
     }
 
 
-    include(editorApplyTestFactory { editor })
+    include(editorApplyTests { editor })
 
     include(
-        editorFieldsTestFactory(
+        editorFieldsTests(
             { editor },
             mapOf(
                 "minDateTime" to {
