+++ src/test/kotlin/com/fwdekker/randomness/datetime/DateTimeSchemeTest.kt
@@ -1,10 +1,10 @@
 package com.fwdekker.randomness.datetime
 
 import com.fwdekker.randomness.integer.IntegerScheme
-import com.fwdekker.randomness.schemeSerializationTestFactory
-import com.fwdekker.randomness.stateDeepCopyTestFactory
 import com.fwdekker.randomness.testhelpers.Tags
+import com.fwdekker.randomness.testhelpers.schemeSerializationTestFactory
 import com.fwdekker.randomness.testhelpers.shouldValidateAsBundle
+import com.fwdekker.randomness.testhelpers.stateDeepCopyTestFactory
 import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.row
