+++ src/test/kotlin/com/fwdekker/randomness/integer/IntegerSchemeTest.kt
@@ -3,10 +3,10 @@ package com.fwdekker.randomness.integer
 import com.fwdekker.randomness.affix.AffixDecorator
 import com.fwdekker.randomness.array.ArrayDecorator
 import com.fwdekker.randomness.fixedlength.FixedLengthDecorator
-import com.fwdekker.randomness.schemeSerializationTestFactory
-import com.fwdekker.randomness.stateDeepCopyTestFactory
 import com.fwdekker.randomness.testhelpers.Tags
+import com.fwdekker.randomness.testhelpers.schemeSerializationTestFactory
 import com.fwdekker.randomness.testhelpers.shouldValidateAsBundle
+import com.fwdekker.randomness.testhelpers.stateDeepCopyTestFactory
 import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.row
