+++ src/test/kotlin/com/fwdekker/randomness/template/TemplateActionLoaderTest.kt
@@ -1,11 +1,8 @@
 package com.fwdekker.randomness.template
 
-import com.fwdekker.randomness.testhelpers.Tags
-import com.fwdekker.randomness.testhelpers.afterNonContainer
 import com.fwdekker.randomness.testhelpers.beforeNonContainer
+import com.fwdekker.randomness.testhelpers.setUpBareIdeaFixture
 import com.intellij.openapi.actionSystem.ActionManager
-import com.intellij.testFramework.fixtures.IdeaTestFixture
-import com.intellij.testFramework.fixtures.IdeaTestFixtureFactory
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.beEmpty
 import io.kotest.matchers.collections.shouldHaveAtLeastSize
@@ -18,26 +15,17 @@ import io.kotest.matchers.shouldNot
  * Unit tests for [TemplateActionLoader].
  */
 object TemplateActionLoaderTest : FunSpec({
-    tags(Tags.IDEA_FIXTURE)
-
-
-    lateinit var ideaFixture: IdeaTestFixture
     lateinit var actionManager: ActionManager
     lateinit var templates: MutableList<Template>
     lateinit var loader: TemplateActionLoader
 
 
-    beforeNonContainer {
-        ideaFixture = IdeaTestFixtureFactory.getFixtureFactory().createBareFixture()
-        ideaFixture.setUp()
+    setUpBareIdeaFixture()
 
+    beforeNonContainer {
         actionManager = ActionManager.getInstance()
         templates = mutableListOf()
         loader = TemplateActionLoader { templates }
-    }
-
-    afterNonContainer {
-        ideaFixture.tearDown()
     }
