+++ src/test/kotlin/com/fwdekker/randomness/template/TemplateActionTests.kt
@@ -4,16 +4,12 @@ import com.fwdekker.randomness.Icons
 import com.fwdekker.randomness.TypeIcon
 import com.fwdekker.randomness.testhelpers.DummyScheme
 import com.fwdekker.randomness.testhelpers.TRANSPARENCY
-import com.fwdekker.randomness.testhelpers.Tags
-import com.fwdekker.randomness.testhelpers.afterNonContainer
 import com.fwdekker.randomness.testhelpers.beSameIconAs
-import com.fwdekker.randomness.testhelpers.beforeNonContainer
 import com.fwdekker.randomness.testhelpers.getSouthColor
 import com.fwdekker.randomness.testhelpers.matchBundle
 import com.fwdekker.randomness.testhelpers.render
+import com.fwdekker.randomness.testhelpers.setUpBareIdeaFixture
 import com.fwdekker.randomness.testhelpers.typeIcon
-import com.intellij.testFramework.fixtures.IdeaTestFixture
-import com.intellij.testFramework.fixtures.IdeaTestFixtureFactory
 import com.intellij.util.ui.EmptyIcon
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.row
@@ -30,20 +26,7 @@ import java.awt.Color
  * Unit tests for [TemplateGroupAction].
  */
 object TemplateGroupActionTest : FunSpec({
-    tags(Tags.IDEA_FIXTURE)
-
-
-    lateinit var ideaFixture: IdeaTestFixture
-
-
-    beforeNonContainer {
-        ideaFixture = IdeaTestFixtureFactory.getFixtureFactory().createBareFixture()
-        ideaFixture.setUp()
-    }
-
-    afterNonContainer {
-        ideaFixture.tearDown()
-    }
+    setUpBareIdeaFixture()
 
 
     context("init") {
@@ -140,20 +123,7 @@ object TemplateInsertActionTest : FunSpec({
  * Unit tests for [TemplateSettingsAction].
  */
 object TemplateSettingsActionTest : FunSpec({
-    tags(Tags.IDEA_FIXTURE)
-
-
-    lateinit var ideaFixture: IdeaTestFixture
-
-
-    beforeNonContainer {
-        ideaFixture = IdeaTestFixtureFactory.getFixtureFactory().createBareFixture()
-        ideaFixture.setUp()
-    }
-
-    afterNonContainer {
-        ideaFixture.tearDown()
-    }
+    setUpBareIdeaFixture()
 
 
     context("init") {
